Module minimal_lexical::bigint

A simple big-integer type for slow path algorithms.

This includes minimal stackvector for use in big-integer arithmetic.

Structs

Bigint

Storage for a big integer type.

ReverseView

REVERSE VIEW Reverse, immutable view of a sequence.

Constants

BIGINT_BITS

Number of bits in a Bigint.

BIGINT_LIMBS

The number of limbs for the bigint.

LIMB_BITS

Functions

bit_length

Calculate the bit-length of the big-integer.

compare

Compare x to y, in little-endian order.

from_u64

Create StackVec from u64 value.

hi64

Get the high 64 bits from the vector.

is_normalized

Get if the big integer is normalized.

large_add

Add bigint to bigint.

large_add_from

Add bigint to bigint starting from offset.

large_mul

Multiply bigint by bigint using grade-school multiplication algorithm.

leading_zeros

Get number of leading zero bits in the storage.

long_mul

Grade-school multiplication algorithm.

nonzero

Check if any of the remaining bits are non-zero.

normalize

Normalize the integer, so any leading zero values are removed.

pow

MulAssign by a power of 5.

rview

Create a reverse view of the vector for indexing.

scalar_add

Add two small integers and return the resulting value and if overflow happens.

scalar_mul

Multiply two small integers (with carry) (and return the overflow contribution).

shl

Shift-left buffer by n bits.

shl_bits

Shift-left n bits inside a buffer.

shl_limbs

Shift-left n limbs inside a buffer.

small_add

Add small integer to bigint.

small_add_from

Add small integer to bigint starting from offset.

small_mul

Multiply bigint by small integer.

u32_to_hi64_1

Shift 32-bit integer to high 64-bits.

u32_to_hi64_2

Shift 2 32-bit integers to high 64-bits.

u32_to_hi64_3

Shift 3 32-bit integers to high 64-bits.

u64_to_hi64_1

Shift 64-bit integer to high 64-bits.

u64_to_hi64_2

Shift 2 64-bit integers to high 64-bits.

Type Definitions

Limb

VecType

Wide