Function minimal_lexical::bigint::long_mul

pub fn long_mul(x: &[Limb], y: &[Limb]) -> Option<VecType>

Grade-school multiplication algorithm.

Slow, naive algorithm, using limb-bit bases and just shifting left for each iteration. This could be optimized with numerous other algorithms, but it’s extremely simple, and works in O(n*m) time, which is fine by me. Each iteration, of which there are m iterations, requires n multiplications, and n additions, or grade-school multiplication.

Don’t use Karatsuba multiplication, since out implementation seems to be slower asymptotically, which is likely just due to the small sizes we deal with here. For example, running on the following data:

const SMALL_X: &[u32] = &[
    766857581, 3588187092, 1583923090, 2204542082, 1564708913, 2695310100, 3676050286,
    1022770393, 468044626, 446028186
];
const SMALL_Y: &[u32] = &[
    3945492125, 3250752032, 1282554898, 1708742809, 1131807209, 3171663979, 1353276095,
    1678845844, 2373924447, 3640713171
];
const LARGE_X: &[u32] = &[
    3647536243, 2836434412, 2154401029, 1297917894, 137240595, 790694805, 2260404854,
    3872698172, 690585094, 99641546, 3510774932, 1672049983, 2313458559, 2017623719,
    638180197, 1140936565, 1787190494, 1797420655, 14113450, 2350476485, 3052941684,
    1993594787, 2901001571, 4156930025, 1248016552, 848099908, 2660577483, 4030871206,
    692169593, 2835966319, 1781364505, 4266390061, 1813581655, 4210899844, 2137005290,
    2346701569, 3715571980, 3386325356, 1251725092, 2267270902, 474686922, 2712200426,
    197581715, 3087636290, 1379224439, 1258285015, 3230794403, 2759309199, 1494932094,
    326310242
];
const LARGE_Y: &[u32] = &[
    1574249566, 868970575, 76716509, 3198027972, 1541766986, 1095120699, 3891610505,
    2322545818, 1677345138, 865101357, 2650232883, 2831881215, 3985005565, 2294283760,
    3468161605, 393539559, 3665153349, 1494067812, 106699483, 2596454134, 797235106,
    705031740, 1209732933, 2732145769, 4122429072, 141002534, 790195010, 4014829800,
    1303930792, 3649568494, 308065964, 1233648836, 2807326116, 79326486, 1262500691,
    621809229, 2258109428, 3819258501, 171115668, 1139491184, 2979680603, 1333372297,
    1657496603, 2790845317, 4090236532, 4220374789, 601876604, 1828177209, 2372228171,
    2247372529
];

We get the following results:

mul/small:long          time:   [220.23 ns 221.47 ns 222.81 ns]
Found 4 outliers among 100 measurements (4.00%)
  2 (2.00%) high mild
  2 (2.00%) high severe
mul/small:karatsuba     time:   [233.88 ns 234.63 ns 235.44 ns]
Found 11 outliers among 100 measurements (11.00%)
  8 (8.00%) high mild
  3 (3.00%) high severe
mul/large:long          time:   [1.9365 us 1.9455 us 1.9558 us]
Found 12 outliers among 100 measurements (12.00%)
  7 (7.00%) high mild
  5 (5.00%) high severe
mul/large:karatsuba     time:   [4.4250 us 4.4515 us 4.4812 us]

In short, Karatsuba multiplication is never worthwhile for out use-case.