Function minimal_lexical::bigint::pow

pub fn pow(x: &mut VecType, exp: u32) -> Option<()>

MulAssign by a power of 5.

Theoretically…

Use an exponentiation by squaring method, since it reduces the time complexity of the multiplication to ~O(log(n)) for the squaring, and O(n*m) for the result. Since m is typically a lower-order factor, this significantly reduces the number of multiplications we need to do. Iteratively multiplying by small powers follows the nth triangular number series, which scales as O(p^2), but where p is n+m. In short, it scales very poorly.

Practically….

Exponentiation by Squaring: running 2 tests test bigcomp_f32_lexical … bench: 1,018 ns/iter (+/- 78) test bigcomp_f64_lexical … bench: 3,639 ns/iter (+/- 1,007)

Exponentiation by Iterative Small Powers: running 2 tests test bigcomp_f32_lexical … bench: 518 ns/iter (+/- 31) test bigcomp_f64_lexical … bench: 583 ns/iter (+/- 47)

Exponentiation by Iterative Large Powers (of 2): running 2 tests test bigcomp_f32_lexical … bench: 671 ns/iter (+/- 31) test bigcomp_f64_lexical … bench: 1,394 ns/iter (+/- 47)

The following benchmarks were run on 1 * 5^300, using native pow, a version with only small powers, and one with pre-computed powers of 5^(3 * max_exp), rather than 5^(5 * max_exp).

However, using large powers is crucial for good performance for higher powers. pow/default time: [426.20 ns 427.96 ns 429.89 ns] pow/small time: [2.9270 us 2.9411 us 2.9565 us] pow/large:3 time: [838.51 ns 842.21 ns 846.27 ns]

Even using worst-case scenarios, exponentiation by squaring is significantly slower for our workloads. Just multiply by small powers, in simple cases, and use precalculated large powers in other cases.

Furthermore, using sufficiently big large powers is also crucial for performance. This is a tradeoff of binary size and performance, and using a single value at ~5^(5 * max_exp) seems optimal.