Struct minimal_lexical::bigint::Bigint

pub struct Bigint {
    pub data: VecType,
}

Storage for a big integer type.

This is used for algorithms when we have a finite number of digits. Specifically, it stores all the significant digits scaled to the proper exponent, as an integral type, and then directly compares these digits.

This requires us to store the number of significant bits, plus the number of exponent bits (required) since we scale everything to the same exponent.

Fields
data: VecType

Significant digits for the float, stored in a big integer in LE order.

This is pretty much the same number of digits for any radix, since the significant digits balances out the zeros from the exponent: 1. Decimal is 1091 digits, 767 mantissa digits + 324 exponent zeros. 2. Base 6 is 1097 digits, or 680 mantissa digits + 417 exponent zeros. 3. Base 36 is 1086 digits, or 877 mantissa digits + 209 exponent zeros.

However, the number of bytes required is larger for large radixes: for decimal, we need log2(10**1091) ≅ 3600, while for base 36 we need log2(36**1086) ≅ 5600. Since we use uninitialized data, we avoid a major performance hit from the large buffer size.

Implementations

impl Bigint

pub fn new() -> Self

Construct a bigint representing 0.

pub fn from_u64(value: u64) -> Self

Construct a bigint from an integer.

pub fn hi64(&self) -> (u64, bool)

pub fn pow(&mut self, base: u32, exp: u32) -> Option<()>

Multiply and assign as if by exponentiation by a power.

pub fn bit_length(&self) -> u32

Calculate the bit-length of the big-integer.